Added a name configuration to indicate which voc to build
In order to know which ontology to build from the configuration file it is
necessary to state which name it has. For this purpose I have added the --name
parameter option to this script (mandatory). The name is both expected to appear
in the configuration file and to be unique.