Added support for external terms
External terms are supported.

Moreover in the case of a term that uses a URI with protocol http but the rest
of the URI is consistent with the URI of this ontology then that URI will not be
considered to be external.

Ie. if this vocabulary has as URI https://example.com/vocabulary then a term
such as http://example.com/vocabulary/terms/Label would not be considered to be
external as opposed to skos:prefLabel.